Remove a loop in UnitOfWork::executeInserts, that caused postPersist events to be called nbDocuments^2 times
I noticed a huge CPU usage when batch inserting 500 documents.
Using xdebug and kcachegrind, I noticed that my PostPersist events were
called 25.000 times. Which is the number of documents, squarred.
This change fixes that bug. It may break things, as I don't understand
why it was done this way. But it doesn't add any failure when I
run the test suite.