Fixing a bug where certain documents are improperly flushed after retrieving the document and making no changes to it.
This occurs when a document has an embedded document that uses a discriminator map. In this case, after retrieving the object, the Hydrator::hydrate() method will register the "original values" for that object (via UnitOfWork::registerManaged()) WITH the discriminator field and value. On the next flush, the presence of the descriminator field and value will make the document appear that it has been modified (i.e. the changeset for the document shows that the original value has the discriminator field and the actual values do not).

The fix is to remove the discriminator field during hydration (after it's used). This is consistent with how discriminator fields are used on non-embedded documents (the discriminator field is removed in UnitOfWorkgetOrCreateDocument()). We don't remove the discriminator field, however, if it's a real, persisted property on the embedded document.