Fixing test MODM48Test.
Because embedded documents are properly given null values, the persistence needs to be smart to treat an entirely empty embedded document as a "new" value (instead of as an update).

Without this change, test MODM48Test throws the following exception "MongoCursorException: can't append to array using string field name [val]".