Add FlushTest
Given 3 users, userA userB userC
userA has a relation to userB
userB has a relation to userC

With only userA in the uow,
If I flush I then have both userA and userB in the uow.
If I flush again, I then have all three users in the uow.

Every flush fetches and registers the relations of the objects in the uow.

Imagine a website where each user has 100 friends.
With one user in the uow, do 2 flushes.
You know have 100*100 = 10,000 users in the uow.