Bug fix in PersistentCollection handling during merge operation.
After serializing/deserializing a document with PersistentCollection(s), initialized collections lose their reference to the DocumentManager and UnitOfWork.
And, as a result, an error will be triggered when trying to clear the collection or do any other operation requiring access to one of these 2 objects.

Another issue is the fact that any collection in a merged document may have different contents than the coresponding document field from the database, so a dirty check is required in order to properly synchronize.