[Mapping] Fix handling of discriminatorField's "name" option
The "name" and "fieldName" options for a discriminator field denote the PHP property and Mongo property for the field, respectively. Based on ClassMetadataInfo::setDiscriminatorField(), "name" is optional and defaults to "fieldName" if not provided.

The annotation driver previously ignored "name" altogether. The XML and YAML drivers required it. This commit fixes those drivers and also adds tests for all four drivers (PHP being the fourth).