Improve persistence of nested collections within embedded documents
Aside from general refactoring, behavior of nested collections has changed. Previously, nested collections were only included if the strategy was not "set"; however, keys were used as-is, which might have caused gaps and resulted in a BSON object being stored in lieu of an array. This fix only ignores nested collections if they are already scheduled for deletion/update (decided in UoW::computeAssociationChanges()). For "pushAll" and "setArray" strategies, a collection to be stored will also first be numerically reindexed to avoid gaps.