Updating documentation to fix distance issue
When using the Coordinate class with properties of latitude and longitude, it appears that latitude always gets taken as the first value in the coordinate system.

Using x and y corrects the order to match MongoDB's near which needs x (longitude) followed by y (latitude). I've also corrected the order of longitude and latitude that was mentioned in the docs.

This was discovered when the distances in my queries were being returned incorrectly due to the problem described above. Leading to many hours of frustration, trust me!