Ignore MongoId constructor exceptions in IdType
Driver 1.4+ will throw an exception if the MongoId constructor argument is invalid. Preserve existing behavior by ignoring this exception and returning a new MongoId.

Note: while this exception makes sense for write operations (the auto identifier mapping should only be used for ObjectId values), it would be undesirable to propogate it during query preparation. In the latter case, one would expect a query on ID for a non-ObjectId string (e.g. "foo") to return an empty result, not yield an exception because the string could not be converted to a MongoId instance.

See: https://jira.mongodb.org/browse/PHP-554