Do not schedule documents with inconsistent IDs for upsert
Documents using IdType that have identifiers invalid for a MongoId constructor were previously upserted with a generated MongoId, but that generated value was never set on the flushed document. As such, it was impossible to immediately query for such documents after flushing. By refusing to schedule such documents for upserting, the insert logic will ensure the generated MongoId is set on the document and the identity map remains consistent.

This change should not disturb handling of documents with alternative ID types (e.g. CustomIdType, IntIdType).