Do not recursively prepare $ne arguments in queries
This test was expecting the array argument for $ne to be prepared recursively. We should just consider its argument as a single value to prepare.

Note that the tests were invalid use cases. The "groups" reference-many association would hold an array of DBRefs. Comparing an array with an ObjectId against "groups.$id" (after translation) would never yield a match, so the negation would always match.