Fix cascade persist for associations.
During a cascade persist operation on a Many association, a snapshot of
the associated collection is taken, marking the collection as
"not dirty".  So if items were added to the collection before persisting,
and the collection was never initialized, the items are lost (persisted
in DB but not present in the instanciated collection anymore).

By using the snapshot during the collection initialization, we can
re-add the lost items after loading the collection.