[WIP] Investigating segfault with requireIndexes
PHP 5.5.12-2ubuntu4.2 (cli) and PHP driver 1.6.4. This appears to be a PHP issue, rather than the driver, as it does not reproduce on 5.6. This seems connected to Collection::isFieldIndexed() fetching the collection indexes on each call. While that is inefficient and should probably be revised, a segfault is still not acceptable.