Refactor creation of managed copy in doMerge
This fixes merge for embedded documents that contain an identifier with a NONE strategy or a custom strategy that requires additional input. The distinction between tryGetById and find is unnecessary since find will always check the identityMap to see if the document has already been loaded to avoid unnecessary database hits.