Remove default value from type attribute
The XSD defines a default value (`string`) for the `type` attribute of the `field` tag. PhpStorm (or an extension?) even issues a warning ("Redundant default attribute value assignment") when I set the attribute to `string`.

So either the code has to be changed to reflect the XML definition, or the definition needs to be changed - which this PR does.