Fix upsert behavior of nullable fields
This brings consistency to the upsert handling of fields with a null values. Non-nullable fields were not unset from the database to prevent loss of data, but nullable fields had their null value set in the daatabase. Since the object being upserted may be an incomplete object that doesn't completely represent the database state, we only write null values to the database if the upsert results in a document insertion.