Fix adding discriminator map to queries
Previously, the discriminator was only added to the query if the class itself had a discriminator value. This is wrong and leads to issues when having multiple levels of document inheritance. In that case, we need to make sure to also include discriminator values if a mapped superclass without discriminator value is found somewhere in the class inheritance chain.