Run appropriate query operations when replacing documents
This ensures that when not using update operators, the query does not attempt to run an updateOne or findOneAndUpdate operation. Instead, the replaceOne or findOneAndReplace operation must be run to avoid errors.
When using the "multiple" option on an update query in combination with a replacement document, the query class will now throw an error as there is no replaceMany functionality.