Proposal: react to entity instance in subscriber
Before I work on this PR I would like to ask on your opinion. I find the way the subscriber is written misleading. Currently, the subsriber will try to link the ODM Document to *every* ORM Entity. With other words the subscriber is  called post load of *every* entity defined in the project. What we want is something like

```
public function postLoad(LifecycleEventArgs $eventArgs)
    {
        $entity = $eventArgs->getEntity();

        if ($entity instanceof Order) {
            // business here
        }
}
```

Am I correct or am I missing a piece of information?

Also, in this PR I would add an example of a .yaml configuration of the listener.