Fix updates to embedded documents after removal
After a parent document and its embeds are removed and then subsequently
persisted, the embedded documents didn't get updated. This happened
because the parent document goes through an upsert operation, but
`PersistenceBuilder->prepareUpsertData()` did not handle embedded
associations in the same way that
`PersistenceBuilder->prepareUpdateData()` does. This fixes that issue so
that an upserted parent document handles embed-many's the same was as an
updated parent document.

Fixes doctrine#2767